Langchain-v2/pyproject.toml
[project]
name = "langchain-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.15",
    "langchain-community>=0.3.15",
    "langchain-google-genai>=2.0.9",
    "langgraph>=0.2.65",
    "langserve[all]>=0.3.1",
    "nbstripout>=0.8.1",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
